fix: rename util.ts
