+++ src/lib/capitalize-first-letter.ts
@@ -1,0 +1,1 @@
+export const capitalizeFirstLetter = (raw: string) => `${raw.charAt(0).toUpperCase()}${raw.slice(1)}`;
